Update aqs scap targets with new hosts and tidy up
We have discovered that the most recent deployment of aqs missed several
hosts, so our new knowledge gaps endpoint is returning errors from
several backend servers.

This commit aims to rectify that by adding the new aqs hosts to the list
of scap targets. This includes aqs10[16-21] as well as all 12 new
servers in codfw aqs20[01-12].

The environment is renamed back from aqs-next to aqs and we also remove
an old reference to a default environment that targets hosts that no
longer exist.

Bug: T342213
Change-Id: I8def9c7241726a8c52d5437a99d6c09d5f5780ee